django-SashaSouer/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.9.3
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configurate==0.4.1
cryptography==41.0.3
cvxpy==1.3.2
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
ecos==2.0.12
exceptiongroup==1.1.3
fake-useragent==1.2.1
fake_user_agent==2.1.8
Flask==2.3.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
lxml==4.9.3
mapping==0.1.6
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
osqp==0.6.3
outcome==1.2.0
pandas==2.1.0
pdfkit==1.0.0
phonenumbers==8.12.25
pika==1.3.2
pika-stubs==0.1.3
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qdldl==0.1.7.post0
requests==2.31.0
requests-oauthlib==1.3.1
scipy==1.11.2
scs==3.2.3
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stem==1.8.2
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.0.4
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
wsproto==1.2.0
yarl==1.9.2

pip~=23.2.1
wheel~=0.41.2
setuptools~=68.1.2
django-filter~=23.2